moved the RenderDoc API header (renderdoc_app.h) to a more centralized location
